+++ R/fitLMMs.R
@@ -2,6 +2,7 @@
 #' @description The PI is used as outcome variable in a linear (mixed) model,
 #' with design variables as regressors. Separate models are fitted for every combination of gene
 #' and PI. fitLMMsSingle() is the workhorse function for a single point pattern,
+#' fitSingleLmmModel() for a single feature in a single point pattern.
 #'
 #' @param obj The result object, from a call to \link{estPis} of \link{addWeightFunction}
 #' @param pis Optional, the pis required. Defaults to all pis in the object
@@ -17,10 +18,6 @@
 #' point patterns. See details.
 #' @param features The features for which to fit linear mixed models.
 #' Defaults to all features in the object
-#' @param moranFormula Formula for Moran's I model fitting
-#' @param addMoransI A boolean, include Moran's I of the cell-wise PIs in the calculation
-#' @param numNNs An integer, the number of nearest neighbours in the
-#' weight matrix for the calculation of the Moran's I statistic
 #' @param ... Passed onto fitLMMsSingle
 #' @details Genes or gene pairs with insufficient observations will be silently
 #' omitted. When randomVars is provided as a vector, independent random
@@ -33,12 +30,9 @@
 #'  different point patterns are assigned to different random effects. Set
 #'  'randomNested' to FALSE to override this behaviour.
 #'
-#'  The Moran's I statistic is used to test whether cell-wise PIs ("nnCell", "nnCellPair", "edge" and "centroid") 
-#'  are spatially autocorrelated across the images. The numeric value of the PI is assigned to the 
-#'  centroid location, and then Moran's I is calculated with a fixed number of numNNs nearest neighbours with equal weights.
 #' @return For fitLMMs(), a list of fitted objects
 #' @export
-#' @seealso \link{buildMoransIDataFrame}, \link{buildDataFrame}
+#' @seealso \link{buildDataFrame}
 #' @order 1
 #' @examples
 #' example(addWeightFunction, "smoppix")
@@ -47,37 +41,26 @@
 #' head(res)
 fitLMMs <- function(
     obj, pis = obj$pis, fixedVars = NULL, randomVars = NULL, verbose = TRUE,
-    returnModels = FALSE, Formula = NULL, randomNested = TRUE, features = getEstFeatures(obj),
-    moranFormula = NULL, addMoransI = FALSE, numNNs = 10, ...) {
-  stopifnot(
-    is.logical(returnModels), is.logical(randomNested), is.logical(addMoransI),
-    is.numeric(numNNs)
-  )
-  if (addMoransI) {
-    weightMats <- lapply(getHypFrame(obj)$centroids, buildMoransIWeightMat, numNNs = numNNs)
-    names(weightMats) <- getHypFrame(obj)$image
-  }
+    returnModels = FALSE, Formula = NULL, randomNested = TRUE, features = getEstFeatures(obj), ...) {
+  stopifnot(is.logical(returnModels), is.logical(randomNested))
   out <- lapply(pis, function(pi) {
     fitLMMsSingle(obj, pi = pi, verbose = verbose, fixedVars = fixedVars, randomVars = randomVars,
         returnModels = returnModels, Formula = Formula, randomNested = randomNested,
-        features = features, weightMats = weightMats, moranFormula = moranFormula,
-        addMoransI = addMoransI, ...
+        features = features, ...
     )
   })
   names(out) <- pis
   return(out)
 }
-#' @param weightMats A list of weight matrices for Moran's I
 #' @return For fitLMMsSingle(), a list of test results, if requested also the linear models are returned
 #' @importFrom lmerTest lmer
-#' @importFrom stats formula anova
-#' @importFrom lme4 lmerControl .makeCC isSingular
+#' @importFrom stats formula terms model.matrix
+#' @importFrom lme4 lmerControl .makeCC isSingular lFormula mkReTrms findbars
 #' @importFrom methods is
 #' @rdname fitLMMs
 #' @order 2
-fitLMMsSingle <- function(
-    obj, pi, fixedVars, randomVars, verbose, returnModels,
-    Formula, randomNested, features, addMoransI, weightMats, moranFormula) {
+fitLMMsSingle <- function(obj, pi, fixedVars, randomVars, verbose, returnModels,
+    Formula, randomNested, features) {
   pi <- match.arg(pi, choices = c(
     "nn", "nnPair", "edge", "centroid", "nnCell", "nnPairCell"))
   foo <- checkPi(obj, pi)
@@ -116,21 +99,6 @@ fitLMMsSingle <- function(
   if (verbose) {
     message("Fitted formula for pi ", pi, ":\n", formChar)
   }
-  if (addMoransI) {
-    moranFormula <- buildFormula(moranFormula,
-                                 fixedVars = morFix <- intersect(
-                                   fixedVars,
-                                   getPPPvars(obj)
-                                 ), randomVars = rvMoran <- intersect(randomVars, getPPPvars(obj)),
-                                 outcome = "MoransI"
-    )
-    MMmoran <- as.logical(length(rvMoran))
-    names(morFix) <- morFix
-    contrastsMoran <- lapply(morFix, function(x) named.contr.sum)
-    if (verbose) {
-      message("Fitted formula for Moran's I for pi ", pi, ":\n", formCharMoran <- characterFormula(moranFormula))
-    }
-  }
   Control <- lmerControl(
     check.conv.grad = .makeCC("ignore", tol = 0.002, relTol = NULL),
     check.conv.singular = .makeCC(action = "ignore", tol = formals(isSingular)$tol),
@@ -150,11 +118,8 @@ fitLMMsSingle <- function(
       vapply(obj$hypFrame$tabObs, FUN.VALUE = double(1), function(x) x[gene])
     })
   }
-  featIds <- (if (is.matrix(tmp)) {
-    colSums(tmp, na.rm = TRUE)
-  } else {
-    tmp
-  }) >= 1
+  featIds <- (if (is.matrix(tmp)) {colSums(tmp, na.rm = TRUE)
+    } else {tmp}) >= 1
   Features <- features[featIds]
   pppDf <- centerNumeric(as.data.frame(obj$hypFrame[, c("image", getPPPvars(obj))]))
   if (is.null(fixedVars)) {
@@ -165,46 +130,133 @@ fitLMMsSingle <- function(
     contrasts <- lapply(discreteVars, function(x) named.contr.sum)
   }
   if(!windowId){
-    prepMat <- prepareMatrix(obj, pi = pi, features = Features)
+    prepMatorList <- prepareMatrixOrList(obj, pi = pi, features = Features)
   }
-  if(cellId){
-    prepTabs <- lapply(obj$hypFrame$ppp, function(x){
-      tab <- table(marks(x)[, c("gene", "cell")])
-      tab[, colnames(tab)!= "NA"]
-    })
-    prepCells <- lapply(seq_along(prepTabs), function(n){
-      eventMarks <- marks(obj$hypFrame$ppp[[n]], drop = FALSE)[, getEventVars(obj), drop = FALSE]
-      eventMarks[match(colnames(prepTabs[[n]]), eventMarks$cell),]
-    })
-  }
-  models <- loadBalanceBplapply(Features, function(gene) {
-    df <- buildDataFrame(obj, gene = gene, pi = pi, pppDf = pppDf, 
-                         prepMat = prepMat, prepTabs = prepTabs, prepCells = prepCells)
-    out <- if (is.null(df) || sum(!is.na(df$pi)) < 3) {
-      NULL
-    } else {
-      moranMod <- if (addMoransI) {
-        finalDf <- buildDataFrame(piMat = df, gene = gene, pi = pi,
-          moransI = TRUE, obj = obj, weightMats = weightMats)
-        W <- 1 / finalDf$Variance
-        fitPiModel(moranFormula, finalDf, contrastsMoran, Control,
-                   MM = MMmoran, Weight = W / sum(W, na.rm = TRUE))
-      } # Run this before nesting
-      if (randomNested) {
-        df <- nestRandom(df, randomVarsSplit, intersect(fixedVars, getPPPvars(obj)))
+  if(windowId){
+    models <- loadBalanceBplapply(Features, function(gene) {
+      df <- buildDataFrame(obj, gene = gene, pi = pi, pppDf = pppDf)
+      out <- if (is.null(df) || sum(!is.na(df$pi)) < 3) {
+        NULL
+      } else {
+        if (randomNested) {
+          df <- nestRandom(df, randomVarsSplit, intersect(fixedVars, getPPPvars(obj)))
+        }
+        contrasts <- contrasts[!names(contrasts) %in% vapply(df, FUN.VALUE = TRUE, is.numeric)]
+        fitPiModel(Formula, df, contrasts,
+                            Control, MM = MM, Weight = df$weight)
       }
-      contrasts <- contrasts[!names(contrasts) %in% vapply(df, FUN.VALUE = TRUE, is.numeric)]
-      piMod <- fitPiModel(Formula, df, contrasts,
-                          Control, MM = MM, Weight = df$weight)
-      list(piMod = piMod, moranMod = moranMod)}
-    return(out)
-  })
+      return(out)
+    })
+  } else {
+    if(cellId){
+      prepTableOrList <- lapply(obj$hypFrame$ppp, function(x){
+        tab <- table(marks(x)[, c("gene", "cell")])
+        class(tab) <- "matrix"
+        t(tab[, colnames(tab)!= "NA"])
+      })
+      prepCells <- lapply(seq_along(prepTableOrList), function(n){
+        eventMarks <- marks(obj$hypFrame$ppp[[n]], drop = FALSE)[, getEventVars(obj), drop = FALSE]
+        eventMarks[match(rownames(prepTableOrList[[n]]), eventMarks$cell),]
+      })
+    } else {
+      prepTableOrList <- {
+        singleFeats <- getFeatures(obj)
+        emptyTab <- matrix(NA, nrow = nrow(pppDf), ncol = length(singleFeats), 
+                          dimnames = list(rownames(pppDf), singleFeats))
+        for(i in rownames(pppDf)){
+          emptyTab[i,names(getHypFrame(obj)[[i,"tabObs"]])] <- getHypFrame(obj)[[i,"tabObs"]]
+        }
+        emptyTab
+      }
+    }
+    #Prepare generic dataframe with fixed and random effects, later 
+    # swap in weights and outcome per feature
+    baseDf <- if(cellId){
+      Reduce(f = rbind, lapply(seq_along(prepCells), function(n) {
+        cbind(prepCells[[n]][, setdiff(colnames(prepCells[[n]]), "gene"), drop = FALSE],
+              pppDf[n,setdiff(colnames(pppDf), "cell"), drop = FALSE])
+      }))
+    } else {
+      pppDf
+    }
+    contrasts <- contrasts[!names(contrasts) %in% vapply(baseDf, FUN.VALUE = TRUE, is.numeric)]
+    if(MM){
+      ff <- lFormula(Formula, data = data.frame("pi" = 0.5, baseDf), 
+                   contrasts = contrasts, na.action = na.omit)
+      if (randomNested) {
+        ff$fr <- nestRandom(ff$fr, randomVarsSplit, intersect(fixedVars, getPPPvars(obj)))
+        ff$reTrms <- mkReTrms(findbars(Formula[[length(Formula)]]), ff$fr)
+      }
+    }
+    modMat <- model.matrix(formula(paste("~",  if(is.null(fixedVars)){"1"} else {paste(collapse = "+", fixedVars)})),
+                          baseDf, contrasts.arg = contrasts) #Fixed effects model matrix
+    Assign <- attr(modMat, "assign")
+    models <- loadBalanceBplapply(Features, function(gene) {
+      mat <- getPiAndWeights(obj, gene = gene, pi = pi, prepMat = prepMatorList,
+                            prepTab = prepTableOrList)
+      out <- if (is.null(mat) || sum(id <- !is.na(mat[, "pi"])) < 3) {
+        NULL
+      } else {
+        if(MM){
+          ff$fr <- ff$fr[id,, drop = FALSE];ff$X <- ff$X[id,, drop = FALSE]
+          attr(ff$X, "assign") <- Assign
+          ff$reTrms$Zt <- ff$reTrms$Zt[, id, drop = FALSE]
+        }
+        fitSingleLmmModel(ff = ff, y = mat[id, "pi"], Terms = terms(Formula), modMat = modMat[id,,drop = FALSE],
+                    weights = mat[id, "weights"], Control = Control, MM = MM, Assign = Assign)
+        }
+      return(out)
+    })
+  }
   names(models) <- Features
-  mods <- c("piMod", if (addMoransI) "moranMod")
-  names(mods) <- mods
-  results <- lapply(mods, function(mm) {
-    extractResults(models, hypFrame = obj$hypFrame, fixedVars, subSet = mm)
-  })
+  results <- extractResults(models, hypFrame = obj$hypFrame, fixedVars)
   # Effect size, standard error, p-value and adjusted p-value per mixed effect
-  return(list(results = results$piMod, resultsMoran = results$moranMod, models = if(returnModels) models))
+  return(list(results = results, models = if(returnModels) models))
+}
+#' Take an existing frame, add outcome and weight and fit lmer model
+#'
+#' @param ff The prepared frame
+#' @param y outcome vector
+#' @param weights weights vector
+#' @param Assign,Terms Added to fitted fixed effects model
+#' @param modMat Design matrix of the fixed effects model
+#' @inheritParams fitPiModel
+#'
+#' @returns A fitted lmer model
+#' @importFrom lme4 mkLmerDevfun optimizeLmer mkMerMod
+#' @importFrom stats lm.wfit
+fitSingleLmmModel <- function(ff, y, Control, Terms, modMat, MM, Assign, weights = NULL) {
+  if(MM){
+    fr <- ff$fr                    # this is a data.frame (model frame)
+    ## Use model-frame column names used by stats::model.frame
+    fr$`(weights)` <- weights
+    fr[["pi - 0.5"]] <- y - 0.5               # replace response
+    mod <- try({
+      devfun <- mkLmerDevfun(fr, ff$X, ff$reTrms, control = Control)
+      opt <- optimizeLmer(devfun, control = Control)
+      out <- mkMerMod(rho = environment(devfun), opt = opt, 
+                      reTrms = ff$reTrms, fr = fr)
+      out <- lmerTest:::as_lmerModLT(out, devfun = devfun)
+    }, silent = TRUE)
+  }
+  # Switch to fixed effects model when fit failed
+  if(!MM || inherits(mod, "try-error")){
+    mod <- lm_from_wfit(lm.wfit(y = y, x = modMat, w = weights), y = y, 
+                        Assign = Assign, Terms = Terms)
+  }
+  return(mod)
+}
+#' Add compoments to a result from lm.wfit to make it a minimally valid lm object
+#'
+#' @param obj The lm.wfit() result
+#' @param y the outcome variable
+#' @param Terms,Assign Added to the object
+#'
+#' @returns A object of class lm
+lm_from_wfit <- function(obj, y, Terms, Assign) {
+  # Create a minimal lm object
+  obj <- c(obj, list(y = y, terms = Terms))
+  obj$assign <- Assign
+  class(obj) <- "lm"
+  return(obj)
 }