+++ R/loadBalanceBplapply.R
@@ -10,7 +10,10 @@
 #' @return A list with the same length as iterator
 #' @importFrom BiocParallel bpparam bplapply bpnworkers
 #' @export
-loadBalanceBplapply <- function(iterator, func, loopFun = "bplapply"){
+#' @examples
+#' library(BiocParallel)
+#' loadBalanceBplapply(LETTERS, length)
+loadBalanceBplapply <- function(iterator, func, loopFun = if(bpparam()$workers==1) "lapply" else "bplapply"){
     loopFunMatched <- match.fun(loopFun)
     if (loopFun == "lapply") {
         loopFunMatched(iterator, func)
@@ -20,12 +23,11 @@ loadBalanceBplapply <- function(iterator, func, loopFun = "bplapply"){
         out <- unsplit(f = splitFac, loopFunMatched(
             split(iterator, f = splitFac),
             function(ss) {
-                out <- lapply(ss, func)
-                names(out) <- ss
+                out <- lapply(selfName(ss), func)
                 return(out)
             }
         ))
-        names(out) = names(iterator)
+        names(out) <- names(iterator)
         return(out)
     }
 }