+++ tests/testthat/test-helpers.R
@@ -9,15 +9,12 @@ test_that("Make gene pairs works as expected", {
     ))
 })
 test_that("getGp does correct extraction", {
-    geneMat <- matrix(rnorm(6), 3, 2)
-    rownames(geneMat) <- makePairs(paste0("gene", seq_len(3)))
+    geneMat <- matrix(rnorm(9), 3, 3)
+    colnames(geneMat) <- makePairs(paste0("gene", seq_len(3)))
     expect_identical(getGp(geneMat, "gene1--gene2"), getGp(geneMat, "gene2--gene1"))
     expect_is(getGp(geneMat, "gene1--gene2", drop = FALSE), "matrix")
 })
 library(lme4)
-test_that("Only fixed part of formula is extracted", {
-    expect_equal(nobars(formula("outcome~fixed + (1|random)")), formula("outcome ~ fixed"))
-})
 test_that("Design variables are well constructed", {
     expect_identical(makeDesignVar(data.frame(a = 1:3, b = LETTERS[1:3]), c(
         "a",